backend/requirements.txt
# fast api
fastapi
uvicorn
# mongodb connector
pymongo
# mariadb connector
mysql-connector-python
# authentication
pyjwt
bcrypt
redis
python-multipart
# env
python-dotenv
# app
numpy
openai